Expose StructureReader to the API
Move TestStructureReader to API Module to fix Test Errors
