+++ api/src/main/java/me/gamercoder215/starcosmetics/api/TestStarConfig.java
@@ -1,6 +1,11 @@
 package me.gamercoder215.starcosmetics.api;
 
+import me.gamercoder215.starcosmetics.api.cosmetics.structure.StructureReader;
 import org.jetbrains.annotations.NotNull;
+
+import java.io.File;
+import java.io.InputStream;
+import java.io.Reader;
 
 final class TestStarConfig implements StarConfig {
 
@@ -43,5 +48,20 @@ final class TestStarConfig implements StarConfig {
 
     @Override
     public void setBlockDisappearTime(long time) throws IllegalArgumentException {}
+
+    @Override
+    public StructureReader getStructureReader(@NotNull File file) {
+        return null;
+    }
+
+    @Override
+    public StructureReader getStructureReader(@NotNull InputStream stream) {
+        return null;
+    }
+
+    @Override
+    public StructureReader getStructureReader(@NotNull Reader reader) {
+        return null;
+    }
 
 }
