+++ api/src/test/java/me/gamercoder215/starcosmetics/api/cosmetics/structure/TestStructureReader.java
